requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
celery
requests
python-dotenv